+++ tests/integrations/django/test_basic.py
@@ -511,7 +511,7 @@ def test_does_not_capture_403(sentry_init, client, capture_events, endpoint):
     assert not events
 
 
-def test_middleware_spans(sentry_init, client, capture_events):
+def test_middleware_spans(sentry_init, client, capture_events, render_span_tree):
     sentry_init(
         integrations=[DjangoIntegration()],
         traces_sample_rate=1.0,
@@ -525,26 +525,36 @@ def test_middleware_spans(sentry_init, client, capture_events):
 
     assert message["message"] == "hi"
 
-    for middleware in transaction["spans"]:
-        assert middleware["op"] == "django.middleware"
-
     if DJANGO_VERSION >= (1, 10):
-        reference_value = [
-            "django.contrib.sessions.middleware.SessionMiddleware.__call__",
-            "django.contrib.auth.middleware.AuthenticationMiddleware.__call__",
-            "tests.integrations.django.myapp.settings.TestMiddleware.__call__",
-            "tests.integrations.django.myapp.settings.TestFunctionMiddleware.__call__",
-        ]
-    else:
-        reference_value = [
-            "django.contrib.sessions.middleware.SessionMiddleware.process_request",
-            "django.contrib.auth.middleware.AuthenticationMiddleware.process_request",
-            "tests.integrations.django.myapp.settings.TestMiddleware.process_request",
-            "tests.integrations.django.myapp.settings.TestMiddleware.process_response",
-            "django.contrib.sessions.middleware.SessionMiddleware.process_response",
-        ]
+        assert (
+            render_span_tree(transaction)
+            == """\
+- op="http.server": description=null
+  - op="django.middleware": description="django.contrib.sessions.middleware.SessionMiddleware.__call__"
+    - op="django.middleware": description="django.contrib.auth.middleware.AuthenticationMiddleware.__call__"
+      - op="django.middleware": description="django.middleware.csrf.CsrfViewMiddleware.__call__"
+        - op="django.middleware": description="tests.integrations.django.myapp.settings.TestMiddleware.__call__"
+          - op="django.middleware": description="tests.integrations.django.myapp.settings.TestFunctionMiddleware.__call__"
+            - op="django.middleware": description="django.middleware.csrf.CsrfViewMiddleware.process_view"
+            - op="django.view": description="message"\
+"""
+        )
 
-    assert [t["description"] for t in transaction["spans"]] == reference_value
+    else:
+        assert (
+            render_span_tree(transaction)
+            == """\
+- op="http.server": description=null
+  - op="django.middleware": description="django.contrib.sessions.middleware.SessionMiddleware.process_request"
+  - op="django.middleware": description="django.contrib.auth.middleware.AuthenticationMiddleware.process_request"
+  - op="django.middleware": description="tests.integrations.django.myapp.settings.TestMiddleware.process_request"
+  - op="django.middleware": description="django.middleware.csrf.CsrfViewMiddleware.process_view"
+  - op="django.view": description="message"
+  - op="django.middleware": description="tests.integrations.django.myapp.settings.TestMiddleware.process_response"
+  - op="django.middleware": description="django.middleware.csrf.CsrfViewMiddleware.process_response"
+  - op="django.middleware": description="django.contrib.sessions.middleware.SessionMiddleware.process_response"\
+"""
+        )
 
 
 def test_middleware_spans_disabled(sentry_init, client, capture_events):
@@ -560,3 +570,30 @@ def test_middleware_spans_disabled(sentry_init, client, capture_events):
     assert message["message"] == "hi"
 
     assert not transaction["spans"]
+
+
+def test_csrf(sentry_init, client):
+    """
+    Assert that CSRF view decorator works even with the view wrapped in our own
+    callable.
+    """
+
+    sentry_init(integrations=[DjangoIntegration()])
+
+    content, status, _headers = client.post(reverse("csrf_hello_not_exempt"))
+    assert status.lower() == "403 forbidden"
+
+    content, status, _headers = client.post(reverse("sentryclass_csrf"))
+    assert status.lower() == "403 forbidden"
+
+    content, status, _headers = client.post(reverse("sentryclass"))
+    assert status.lower() == "200 ok"
+    assert b"".join(content) == b"ok"
+
+    content, status, _headers = client.post(reverse("classbased"))
+    assert status.lower() == "200 ok"
+    assert b"".join(content) == b"ok"
+
+    content, status, _headers = client.post(reverse("message"))
+    assert status.lower() == "200 ok"
+    assert b"".join(content) == b"ok"