+++ CHANGES.md
@@ -27,6 +27,27 @@ sentry-sdk==0.10.1
 
 A major release `N` implies the previous release `N-1` will no longer receive updates. We generally do not backport bugfixes to older versions unless they are security relevant. However, feel free to ask for backports of specific commits on the bugtracker.
 
+## 0.17.0
+
+* Fix a bug where class-based callables used as Django views (without using
+  Django's regular class-based views) would not have `csrf_exempt` applied.
+* New integration for Google Cloud Functions.
+* Fix a bug where a recently released version of `urllib3` would cause the SDK
+  to enter an infinite loop on networking and SSL errors.
+* **Breaking change**: Remove the `traceparent_v2` option. The option has been
+  ignored since 0.16.3, just remove it from your code.
+
+## 0.16.5
+
+* Fix a bug that caused Django apps to crash if the view didn't have a `__name__` attribute.
+
+## 0.16.4
+
+* Add experiment to avoid trunchating span descriptions. Initialize with
+  `init(_experiments={"smart_transaction_trimming": True})`.
+* Add a span around the Django view in transactions to distinguish its
+  operations from middleware operations.
+
 ## 0.16.3
 
 * Fix AWS Lambda support for Python 3.8.